mv control_manager to hvac_secondary + consignes to gen_hvac
+ add existence test on temperature_typelimits
