+++ src/idfhub/hvac24_1_0.py
@@ -13,7 +13,7 @@ from idfhub.hvac import (
 )
 
 from idfhub.idf_autocomplete.v24_1_0.idf_helpers_short import (
-    Scheduletypelimits,ScheduleConstant,ScheduleConstantMeta,
+    Scheduletypelimits,ScheduletypelimitsMeta,ScheduleConstant,ScheduleConstantMeta,
     SiteGroundtemperatureBuildingsurface,
     SiteGroundtemperatureUndisturbedKusudaachenbach,
     HeatpumpWatertowaterEquationfitHeating, HeatpumpPlantloopEirHeating,
@@ -25,8 +25,6 @@ from idfhub.idf_autocomplete.v24_1_0.idf_helpers_short import (
     ScheduleCompact,
     Plantequipmentlist, Plantequipmentoperationschemes,
     PlantequipmentoperationHeatingload, PlantequipmentoperationCoolingload,
-    EnergymanagementsystemSensor, EnergymanagementsystemSensorMeta, EnergymanagementsystemActuator,
-    EnergymanagementsystemProgram, EnergymanagementsystemProgramcallingmanager,
     OutdoorairNode,
     CurveBiquadratic, CurveQuadratic, CurveQuadlinear,
     BoilerHotwater,
@@ -45,8 +43,6 @@ from idfhub.idf_autocomplete.v24_1_0.idf_types_short import (
     ScheduleCompactType,
     PlantequipmentlistType, PlantequipmentoperationschemesType,
     PlantequipmentoperationHeatingloadType, PlantequipmentoperationCoolingloadType,
-    EnergymanagementsystemSensorType, EnergymanagementsystemActuatorType,
-    EnergymanagementsystemProgramType, EnergymanagementsystemProgramcallingmanagerType,
     OutdoorairNodeType,
     CurveBiquadraticType, CurveQuadraticType,CurveQuadlinearType,
     BoilerHotwaterType,
@@ -64,105 +60,19 @@ if not idf:
     sys.exit()
 
 
-def create_sensor(*, sensor_name, sensor_type, location_name):
-    """create a sensor"""
-    sensor = idf.getobject(
-        EnergymanagementsystemSensorMeta.idf_name,
-        sensor_name
+temperature_typelimits = idf.getobject(
+        ScheduletypelimitsMeta.idf_name,
+        "temperature"
     )
-    if sensor is None:
-        return EnergymanagementsystemSensor(
-            idf,
-            **EnergymanagementsystemSensorType(
-                Name=sensor_name,
-                OutputVariable_or_OutputMeter_Index_Key_Name=location_name,
-                OutputVariable_or_OutputMeter_Name=sensor_type
-            )
+if not temperature_typelimits:
+    temperature_typelimits = Scheduletypelimits(
+        idf,
+        **ScheduletypelimitsType(
+            Name="temperature",
+            Numeric_Type=EPValues.CONTINUOUS,
+            Unit_Type=EPValues.TEMPERATURE
         )
-    return sensor
-
-
-def control_manager(sensor_name, conf: dict[str, Any]):
-    """manage equipment availability using a sensor"""
-    _node_name = LoopNodes(conf["loop"]).get(
-         side=conf["side"],
-         port=conf["port"]
     )
-    create_sensor(
-        sensor_name = sensor_name,
-        sensor_type = conf["type"],
-        location_name = _node_name,
-    )
-    for equipment_name in conf.get("controls", []):
-        if "pump" not in equipment_name:
-            continue
-        actuator_name = f"{equipment_name}_availability"
-        program_name = f"{equipment_name}_program"
-        schedule_name = f"{equipment_name}_availability_schedule"
-        schedule = ScheduleCompact(
-            idf,
-            **ScheduleCompactType(
-                Name=schedule_name,
-                Schedule_Type_Limits_Name="Fractional",
-                Field_1=f"{EPValues.THROUGH}: 12/31",
-                Field_2=f"{EPValues.FOR}: {EPValues.ALLDAYS}",
-                Field_3=f"{EPValues.UNTIL}: 24:00",
-                Field_4=1
-            )
-        )
-        EnergymanagementsystemActuator(
-            idf,
-            **EnergymanagementsystemActuatorType(
-                Name=actuator_name,
-                Actuated_Component_Unique_Name=schedule.Name,
-                Actuated_Component_Type=schedule.key,
-                Actuated_Component_Control_Type="Schedule Value"
-            )
-        )
-        ems_instructions = []
-        value = conf.get("stop_below", -5)
-        ems_instructions.append(f"IF {sensor_name} < {value}")
-        value = conf.get("min_flow", 0.1)
-        ems_instructions.append(f"SET {actuator_name} = {value}")
-        ems_instructions.append("ENDIF")
-        value = conf.get("start_above", -5)
-        ems_instructions.append(f"IF {sensor_name} >= {value}")
-        value = conf.get("normal_flow")
-        ems_instructions.append(f"SET {actuator_name} = {value}")
-        ems_instructions.append("ENDIF")
-
-        program = EnergymanagementsystemProgram(
-            idf,
-            **EnergymanagementsystemProgramType(
-                Name=program_name)
-        )
-        for i, instruction in enumerate(ems_instructions):
-            program[f"Program_Line_{i+1}"] = instruction
-        EnergymanagementsystemProgramcallingmanager(
-            idf,
-            **EnergymanagementsystemProgramcallingmanagerType(
-                Name=f"{equipment_name}_control",
-                EnergyPlus_Model_Calling_Point="InsideHVACSystemIterationLoop",
-                Program_Name_1=program_name
-            )
-        )
-        equipments[equipment_name]["Pump_Flow_Rate_Schedule_Name"] = schedule.Name
-
-
-#------------------------------------------------------------------------------
-# on crée 2 schedules constants, 20°C chauffage et 25°C raffraichissement :
-# - const_temp_sched_20deg
-# - const_temp_sched_25deg
-#------------------------------------------------------------------------------
-
-temperature_typelimits = Scheduletypelimits(
-    idf,
-    **ScheduletypelimitsType(
-        Name="temperature",
-        Numeric_Type=EPValues.CONTINUOUS,
-        Unit_Type=EPValues.TEMPERATURE
-    )
-)
 
 # on utilise un schedule compact
 # Mots-clés utiles dans For:
@@ -233,9 +143,6 @@ def constant_schedule(
             Hourly_Value=value
         )
     )
-
-consigne_cool = constant_schedule(25)
-consigne_heat = constant_schedule(20)
 
 #------------------------------------------------------------------------------
 # SETPOINTS
